Crate deskulpt_core

This crate implements the core of the Deskulpt backend.

âš ï¸ This crate is meant to be used internally by the Deskulpt application and is not designed to support plugin authors or other users directly. Private items are documented for reference of Deskulpt developers.

Modules

bundler ğŸ”’

Bundler for Deskulpt widgets.

commands

Deskulpt core commands to be invoked by the frontend.

config ğŸ”’

Configuration of Deskulpt widgets.

events ğŸ”’

Event system for IPC between Deskulpt frontend and backend.

path ğŸ”’

Deskulpt path utilities.

settings ğŸ”’

Application and widget settings.

shortcuts ğŸ”’

Keyboard shortcut management.

states ğŸ”’

Deskulpt runtime state management.

tray ğŸ”’

Deskulpt system tray.

window ğŸ”’

Deskulpt windows.

Structs

Settings

Full settings of the application.

Traits

EventsExt

Extension trait for event-related operations.

PathExt

Extension trait for path-related operations.

ShortcutsExt

Extension trait for keyboard shortcuts.

StatesExtCanvasClickThrough

Extension trait for operations on canvas click-through state.

StatesExtInitialRender

Extension trait for operations related to the initial render.

StatesExtWidgetConfigMap

Extension trait for operations on widget configuration map state.

TrayExt

Extention trait for system tray-related operations.

WindowExt

Extention trait for window-related operations.

Functions

on_window_event

Window event handler for all Deskulpt windows.