deskulpt_core

Module commands

Deskulpt core commands to be invoked by the frontend.

Modules

error 🔒

Command errors.

Functions

bundle_widget

Bundle a widget.

call_plugin

Call a plugin command (🚧 TODO 🚧).

emit_on_render_ready

Wrapper of emit_on_render_ready.

exit_app

Exit the application with cleanup.

open_widget

Open the widgets directory or a specific widget directory.

rescan_widgets

Rescan the widgets directory and update the widget configuration map.

set_render_ready

Wrapper of set_render_ready.

update_shortcut

Wrapper of update_shortcut.