deskulpt_core::commands

Function call_plugin

pub async fn call_plugin<R: Runtime>(
    app_handle: AppHandle<R>,
    plugin: String,
    command: String,
    id: String,
    payload: Option<Value>,
) -> Result<Value, CmdError>

Call a plugin command (🚧 TODO 🚧).

🚧 TODO 🚧

The Deskulpt core should keep a state of the registered plugins and call the plugins dynamically. Also, instead of invoking the plugins directly, the Deskulpt core should not depend on any of the plugins and should use IPC to communicate with the plugins.

Also, in order to simplify the engine API for the plugin (because it is a temporary implementation), app_handle is using the default runtime but it should be a generic R: Runtime parameter in the final implementation.