deskulpt_core::config

Trait LoadFromFile

trait LoadFromFile: Sized + DeserializeOwned {
    const FILE_NAME: &'static str;

    // Provided method
    fn load(dir: &Path) -> Result<Option<Self>> { ... }
}

Helper trait for loading configuration files from a directory.

Required Associated Constants

const FILE_NAME: &'static str

The name of the configuration file.

Provided Methods

fn load(dir: &Path) -> Result<Option<Self>>

Load the configuration file from the given directory.

Specially, this method returns Ok(None) if the target file does not exist and does not treat it as an error.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.