deskulpt_core::path

Trait PathExt

pub trait PathExt<R: Runtime>: Manager<R> {
    // Provided methods
    fn init_widgets_dir(&self) -> Result<()> { ... }
    fn widgets_dir(&self) -> Result<&Path> { ... }
    fn init_persist_dir(&self) -> Result<()> { ... }
    fn persist_dir(&self) -> Result<&Path> { ... }
}

Extension trait for path-related operations.

Provided Methods

fn init_widgets_dir(&self) -> Result<()>

Initialize the widgets directory.

This will create the widgets directory if it does not exist. It must be called before calling the widgets_dir method.

fn widgets_dir(&self) -> Result<&Path>

Get a reference to the widgets directory.

This will create the widgets directory if it does not exist, which can happen if one removes that directory during the application. This will error if the init_widgets_dir method has not been called.

fn init_persist_dir(&self) -> Result<()>

Initialize the persistence directory.

This will create the persistence directory if it does not exist. It must be called before calling the persist_dir method.

fn persist_dir(&self) -> Result<&Path>

Get a reference to the persistence directory.

This will create the persistence directory if it does not exist, which can happen if one removes that directory during the application. This will error if the init_persist_dir method has not been called.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.