deskulpt_core::settings::shortcuts

Function reregister_shortcut

fn reregister_shortcut<R: Runtime>(
    gs: &GlobalShortcut<R>,
    key: &ShortcutKey,
    old: Option<&str>,
    new: Option<&str>,
) -> Result<()>

Helper function for re-registering a keyboard shortcut.

If the old and new shortcuts are the same, this is a no-op. Otherwise, the old shortcut will be unregistered and the new shortcut will be registered with the listener determined by the shortcut key.