deskulpt_core::shortcuts

Function reregister_shortcut

fn reregister_shortcut<R: Runtime>(
    gs: &GlobalShortcut<R>,
    action: &ShortcutAction,
    old: Option<&String>,
    new: Option<&String>,
) -> Result<()>

Re-register a shortcut.

The old shortcut will be unregistered and the new shortcut will be registered, with the listener determined by the shortcut action.