deskulpt_core::shortcuts

Macro impl_shortcuts

macro_rules! impl_shortcuts {
    ($($key: ident => $listener: expr),* $(,)?) => { ... };
}

Implement ShortcutKey and ShortcutsExt for the given shortcuts.

This macro takes a list of key => listener pairs, where key corresponds to the keys of Shortcuts and listener is the corresponding shortcut handler callback.