deskulpt_core::states::canvas_click_through

Trait StatesExtCanvasClickThrough

pub trait StatesExtCanvasClickThrough<R: Runtime>: Manager<R> + EventsExt<R> {
    // Provided methods
    fn manage_canvas_click_through(&self) { ... }
    fn set_canvas_click_through_menu_item(&self, menu_item: &MenuItem<R>) { ... }
    fn toggle_canvas_click_through(&self) -> Result<()> { ... }
}

Extension trait for operations on canvas click-through state.

Provided Methods

fn manage_canvas_click_through(&self)

Initialize state management for whether the canvas is click-through.

The canvas is click-through by default.

fn set_canvas_click_through_menu_item(&self, menu_item: &MenuItem<R>)

Set the menu item for toggling the canvas click-through state.

fn toggle_canvas_click_through(&self) -> Result<()>

Toggle the click-through state of the canvas window.

This will also update the menu item text and show a toast message on the canvas if possible.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.