deskulpt_core::states::initial_render

Trait StatesExtInitialRender

pub trait StatesExtInitialRender<R: Runtime>: Manager<R> + EventsExt<R> {
    // Provided methods
    fn manage_initial_render(&self) { ... }
    fn set_render_ready(&self) -> Result<()> { ... }
    fn emit_on_render_ready(&self, payload: Value) -> Result<()> { ... }
}

Extension trait for operations related to the initial render.

Provided Methods

fn manage_initial_render(&self)

Initialize state management for the initial render.

fn set_render_ready(&self) -> Result<()>

Set the canvas as ready to render widgets.

If there is a pending payload, a render-widgets event will be emitted to the canvas with that payload.

fn emit_on_render_ready(&self, payload: Value) -> Result<()>

Emit a render-widgets event to the canvas when it is ready.

If the canvas is already ready to render widgets, emit the given payload to the canvas immediately. Otherwise, store the payload as pending so that it can be emitted later when the canvas is ready.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.