deskulpt_core::states::logging

Trait LoggingStateExt

pub trait LoggingStateExt<R: Runtime>: Manager<R> + PathExt<R> {
    // Provided methods
    fn manage_logging(&self) -> Result<()> { ... }
    fn collect_logs(&self) -> Result<Vec<PathBuf>> { ... }
    fn clear_logs(&self) -> Result<u64> { ... }
}

Extension trait for operations related to logging.

Provided Methods

fn manage_logging(&self) -> Result<()>

Initialize state management for logging.

This will set up structured logging in newline-delimited JSON format with daily rotation and maximum MAX_LOG_FILES log files retained. A panic hook is also set up to log uncaught panics.

fn collect_logs(&self) -> Result<Vec<PathBuf>>

Discover log files and return their paths by newest first.

This looks for log files in the logs directory with names matching the pattern deskulpt.*.log, where * should be a timestamp though this is not verified here. The returned list is sorted by filename in descending order, which should correspond to most recent first if the *s are indeed timestamps.

fn clear_logs(&self) -> Result<u64>

Clear all log files and return the freed disk space in bytes.

The latest log file is truncated instead of deleted to ensure that logging can continue without interruption. All older log files are deleted. The total freed disk space is returned.

Note that failure to delete or truncate a log file will not result in an error, but will not contribute to the computed freed space. Failure to discover the log files in the first place (before actual clearing begins), however, willl result in an error.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.