deskulpt_core::states::settings

Trait SettingsStateExt

pub trait SettingsStateExt<R: Runtime>:
    Manager<R>
    + Emitter<R>
    + PathExt<R>
    + GlobalShortcutExt<R> {
    // Provided methods
    fn manage_settings(&self) { ... }
    fn get_settings(&self) -> RwLockReadGuard<'_, Settings> { ... }
    fn update_settings<F>(&self, update: F) -> Result<()>
       where F: FnOnce(&Settings) -> SettingsPatch,
             Self: Sized { ... }
}

Extension trait for operations on the settings state.

Provided Methods

fn manage_settings(&self)

Initialize state management for the settings.

This will load the settings from the persistence directory and initialize the shortcuts. If any step fails, it will fall back to a state that preserves as much persisted data as possible.

fn get_settings(&self) -> RwLockReadGuard<'_, Settings>

Get an immutable reference to the settings.

The returned reference is behind a lock guard, which should be dropped as soon as possible to minimize critical section.

fn update_settings<F>(&self, update: F) -> Result<()>
where F: FnOnce(&Settings) -> SettingsPatch, Self: Sized,

Update the settings.

The update closure has access to the current settings and is expected to return a SettingsPatch that describes the desired updates. See its documentation for more information of how the patch will be applied. Patch application is best-effort: any part of the patch that fails to be applied will be skipped, and the rest will be applied as normal. An UpdateSettingsEvent will be emitted to notify the frontend of the changes. Errors will be accumulated and returned as a single error at the end if any occurred.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.