deskulpt_core::states::widget_catalog

Trait WidgetCatalogStateExt

pub trait WidgetCatalogStateExt<R: Runtime>: Manager<R> + PathExt<R> {
    // Provided methods
    fn manage_widget_catalog(&self) { ... }
    fn get_widget_catalog(&self) -> RwLockReadGuard<'_, WidgetCatalog> { ... }
    fn get_widget_catalog_mut(&self) -> RwLockWriteGuard<'_, WidgetCatalog> { ... }
}

Extension trait for operations on widget catalog state.

Provided Methods

fn manage_widget_catalog(&self)

Initialize state management for the widget catalog.

fn get_widget_catalog(&self) -> RwLockReadGuard<'_, WidgetCatalog>

Get an immutable reference to the widget catalog.

The returned reference is behind a lock guard, which should be dropped as soon as possible to minimize critical section.

fn get_widget_catalog_mut(&self) -> RwLockWriteGuard<'_, WidgetCatalog>

Get a mutable reference to the widget catalog.

The returned reference is behind a lock guard, which should be dropped as soon as possible to minimize critical section.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.