deskulpt_core::states::widget_config_map

Trait StatesExtWidgetConfigMap

pub trait StatesExtWidgetConfigMap<R: Runtime>: Manager<R> + PathExt<R> {
    // Provided methods
    fn manage_widget_config_map(&self) { ... }
    fn widget_dir<S: AsRef<str>>(&self, id: S) -> Result<PathBuf> { ... }
    fn with_widget_config_map<F, T>(&self, f: F) -> T
       where F: FnOnce(&HashMap<String, WidgetConfig>) -> T { ... }
    fn with_widget_config_map_mut<F, T>(&self, f: F) -> T
       where F: FnOnce(&mut HashMap<String, WidgetConfig>) -> T { ... }
}

Extension trait for operations on widget configuration map state.

Provided Methods

fn manage_widget_config_map(&self)

Initialize state management for the widget configuration map.

fn widget_dir<S: AsRef<str>>(&self, id: S) -> Result<PathBuf>

Get the directory of a widget by ID.

This will error if the widgets directory cannot be accessed or if the widget is not found in the collection.

fn with_widget_config_map<F, T>(&self, f: F) -> T
where F: FnOnce(&HashMap<String, WidgetConfig>) -> T,

Provide reference to the widget configuration map within a closure.

This will lock the widget configuration map state. The return value of the closure will be propagated.

fn with_widget_config_map_mut<F, T>(&self, f: F) -> T
where F: FnOnce(&mut HashMap<String, WidgetConfig>) -> T,

Provide mutable reference to the widget configuration map within a closure.

This will lock the widget configuration map state. The return value of the closure will be propagated.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.