deskulpt_core::states::widgets

Trait WidgetsStateExt

pub trait WidgetsStateExt<R: Runtime>:
    Manager<R>
    + PathExt<R>
    + SettingsStateExt<R> {
    // Provided methods
    fn manage_widgets(&self) { ... }
    fn get_widget_catalog(&self) -> RwLockReadGuard<'_, WidgetCatalog> { ... }
    fn get_widget_catalog_mut(&self) -> RwLockWriteGuard<'_, WidgetCatalog> { ... }
    fn reload_widgets_all(&self) -> Result<()>
       where Self: Sized { ... }
    fn render_widget(&self, id: String) -> Result<()> { ... }
    fn render_widgets_all(&self) -> Result<()> { ... }
}

Extension trait for operations on Deskulpt widgets.

Provided Methods

fn manage_widgets(&self)

Initialize state management for Deskulpt widgets.

fn get_widget_catalog(&self) -> RwLockReadGuard<'_, WidgetCatalog>

Get an immutable reference to the widget catalog.

The returned reference is behind a lock guard, which should be dropped as soon as possible to minimize critical section.

fn get_widget_catalog_mut(&self) -> RwLockWriteGuard<'_, WidgetCatalog>

Get a mutable reference to the widget catalog.

The returned reference is behind a lock guard, which should be dropped as soon as possible to minimize critical section.

fn reload_widgets_all(&self) -> Result<()>
where Self: Sized,

Reload all widgets.

This function loads a new widget catalog from the widgets directory and replaces the existing catalog. It then syncs the settings with the updated catalog. If any step fails, an error is returned.

fn render_widget(&self, id: String) -> Result<()>

Render a specific widget by its ID.

If the widget does not exist in the catalog, an error is returned. Otherwise, a render task is sent to the render worker. If this fails, an error is returned as well. This function does not wait for the rendering to complete.

fn render_widgets_all(&self) -> Result<()>

Render all widgets in the catalog.

This function sends render tasks for all widgets in the catalog to the render worker. If any task fails to be sent, an error is returned containing all individual errors. This function does not wait for the rendering to complete.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.