deskulpt_core::tray

Function on_menu_event

fn on_menu_event<R: Runtime>(app_handle: &AppHandle<R>, event: MenuEvent)

Handler for system tray menu events.

This handler will receive any menu event but only act on events related to the system tray.