Crate deskulpt_core

This crate implements the core of the Deskulpt backend.

âš ï¸ This crate is meant to be used internally by the Deskulpt application and is not designed to support plugin authors or other users directly. Private items are documented for reference of Deskulpt developers.

Modules

commands ğŸ”’

Deskulpt core commands to be invoked by the frontend.

events

Deskulpt core events.

shortcuts

Keyboard shortcut management.

states

Deskulpt runtime state management.

tray

Deskulpt system tray.

window

Deskulpt windows.

Functions

init

Initialize the plugin.