deskulpt_core::states::settings

Trait SettingsStateExt

pub trait SettingsStateExt<R: Runtime>:
    Manager<R>
    + PathExt<R>
    + GlobalShortcutExt<R> {
    // Provided methods
    fn manage_settings(&self) { ... }
    fn get_settings(&self) -> RwLockReadGuard<'_, Settings> { ... }
    fn get_settings_mut(&self) -> RwLockWriteGuard<'_, Settings> { ... }
    fn update_settings_theme(&self, theme: Theme) { ... }
    fn update_settings_shortcut(
        &self,
        key: ShortcutKey,
        shortcut: Option<String>,
    ) -> Result<()> { ... }
    fn update_settings_widget(&self, id: String, update: WidgetSettings) { ... }
}

Extension trait for operations on the settings state.

Provided Methods

fn manage_settings(&self)

Initialize state management for the settings.

This will load the settings from the persistence directory and initialize the shortcuts. If any step fails, it will fall back to a state that preserves as much persisted data as possible.

fn get_settings(&self) -> RwLockReadGuard<'_, Settings>

Get an immutable reference to the settings.

fn get_settings_mut(&self) -> RwLockWriteGuard<'_, Settings>

Get a mutable reference to the settings.

fn update_settings_theme(&self, theme: Theme)

Update the theme.

fn update_settings_shortcut( &self, key: ShortcutKey, shortcut: Option<String>, ) -> Result<()>

Update a shortcut.

The side effect is that the involved shortcuts will be unregistered or registered. If the provided shortcut is None, it means removing that shortcut.

fn update_settings_widget(&self, id: String, update: WidgetSettings)

Update the settings of a widget.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.