deskulpt_core::states::settings

Trait SettingsStateExt

pub trait SettingsStateExt<R: Runtime>:
    Manager<R>
    + Emitter<R>
    + PathExt<R> {
    // Provided methods
    fn manage_settings(&self) { ... }
    fn get_settings(&self) -> RwLockReadGuard<'_, Settings> { ... }
    fn on_theme_change<F>(&self, hook: F)
       where F: Fn(&Theme, &Theme) + Send + Sync + 'static { ... }
    fn on_shortcut_change<F>(&self, hook: F)
       where F: Fn(&ShortcutKey, Option<&String>, Option<&String>) + Send + Sync + 'static { ... }
    fn update_settings<F>(&self, update: F) -> Result<()>
       where F: FnOnce(&Settings) -> SettingsPatch,
             Self: Sized { ... }
}

Extension trait for operations on the settings state.

Provided Methods

fn manage_settings(&self)

Initialize state management for the settings.

This will load the settings from the persistence directory and initialize the shortcuts. If any step fails, it will fall back to a state that preserves as much persisted data as possible.

fn get_settings(&self) -> RwLockReadGuard<'_, Settings>

Get an immutable reference to the settings.

The returned reference is behind a lock guard, which should be dropped as soon as possible to minimize critical section.

fn on_theme_change<F>(&self, hook: F)
where F: Fn(&Theme, &Theme) + Send + Sync + 'static,

Register a hook that will be triggered on theme change.

The two arguments are respectively the old and new themes.

fn on_shortcut_change<F>(&self, hook: F)
where F: Fn(&ShortcutKey, Option<&String>, Option<&String>) + Send + Sync + 'static,

Register a hook that will be triggered on shortcut change.

The first argument is the shortcut key. The second and third arguments are respectively the old and new shortcuts. None means that no shortcut was/is assigned for that key.

fn update_settings<F>(&self, update: F) -> Result<()>
where F: FnOnce(&Settings) -> SettingsPatch, Self: Sized,

Update the settings.

TODO

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.